<Fix> 로그인 페이지 하단 흰색 배경 제거
